Move polymer element descriptor to polymer/ dir
