+++ src/ast/element-descriptor.ts
@@ -25,31 +25,6 @@ import {BehaviorOrName, Descriptor, EventDescriptor, LiteralValue, LocationOffse
 
 export {Visitor} from '../javascript/estree-visitor';
 
-export interface Options {
-  tagName?: string;
-  className?: string;
-  superClass?: string;
-  extends?: string;
-  jsdoc?: jsdoc.Annotation;
-  description?: string;
-  properties?: Property[];
-  attributes?: Attribute[];
-  observers?: {
-    javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
-  }[];
-  behaviors?: BehaviorOrName[];
-
-  demos?: {desc: string; path: string}[];
-  events?: EventDescriptor[];
-
-  domModule?: dom5.Node;
-  scriptElement?: dom5.Node;
-
-  abstract?: boolean;
-  sourceLocation?: SourceLocation;
-}
-
 export interface Attribute {
   name: string;
   sourceLocation: SourceLocation;
@@ -67,17 +42,6 @@ export interface Property {
   readOnly?: boolean;
   javascriptNode?: estree.Node;
   sourceLocation: SourceLocation;
-}
-
-export interface PolymerProperty extends Property {
-  published?: boolean;
-  notify?: boolean;
-  observer?: string;
-  observerNode?: estree.Expression|estree.Pattern;
-  reflectToAttribute?: boolean;
-  configuration?: boolean;
-  getter?: boolean;
-  setter?: boolean;
 }
 
 export class ElementDescriptor implements Descriptor {
@@ -110,64 +74,4 @@ export class ElementDescriptor implements Descriptor {
     }
   }
 }
-/**
- * The metadata for a single polymer element
- */
-export class PolymerElementDescriptor extends ElementDescriptor {
-  properties: PolymerProperty[] = [];
-  observers: {
-    javascriptNode: estree.Expression | estree.SpreadElement,
-    expression: LiteralValue
-  }[] = [];
-  behaviors: string[] = [];
-  domModule?: dom5.Node;
-  scriptElement?: dom5.Node;
 
-  abstract?: boolean;
-
-  constructor(options: Options) {
-    super();
-    Object.assign(this, options);
-  }
-
-  addProperty(prop: PolymerProperty) {
-    if (prop.name.startsWith('_') || prop.name.endsWith('_')) {
-      prop.private = true;
-    }
-    this.properties.push(prop);
-    const attributeName = propertyToAttributeName(prop.name);
-    if (prop.private || !attributeName) {
-      return;
-    }
-    if (!isFunctionDescriptor(prop)) {
-      this.attributes.push({
-        name: attributeName,
-        sourceLocation: prop.sourceLocation,
-        description: prop.description,
-        type: prop.type,
-      });
-    }
-    if (prop.notify) {
-      this.events.push({
-        name: `${attributeName}-changed`,
-        description: `Fired when ${attributeName} changes.`,
-      });
-    }
-  }
-}
-
-
-/**
- * Implements Polymer core's translation of property names to attribute names.
- *
- * Returns null if the property name cannot be so converted.
- */
-function propertyToAttributeName(propertyName: string): string|null {
-  // Polymer core will not map a property name that starts with an uppercase
-  // character onto an attribute.
-  if (propertyName[0].toUpperCase() === propertyName[0]) {
-    return null;
-  }
-  return propertyName.replace(
-      /([A-Z])/g, (_: string, c1: string) => `-${c1.toLowerCase()}`);
-}
