+++ src/polymer/behavior-finder.ts
@@ -15,7 +15,7 @@
 import * as estraverse from 'estraverse';
 import * as estree from 'estree';
 
-import {BehaviorDescriptor, BehaviorOrName, Descriptor, LiteralValue, PolymerElementDescriptor, Property} from '../ast/ast';
+import {BehaviorDescriptor, BehaviorOrName, Descriptor, LiteralValue, Property} from '../ast/ast';
 import * as astValue from '../javascript/ast-value';
 import {Visitor} from '../javascript/estree-visitor';
 import * as esutil from '../javascript/esutil';
@@ -26,6 +26,7 @@ import * as jsdoc from '../javascript/jsdoc';
 import * as analyzeProperties from './analyze-properties';
 import {PropertyHandlers, declarationPropertyHandlers} from './declaration-property-handlers';
 import * as docs from './docs';
+import {PolymerElementDescriptor} from './element-descriptor';
 
 interface KeyFunc<T> {
   (value: T): any;
